+++ phase2/src/usecases/notes/MemoManager.java
@@ -40,10 +40,13 @@ public class MemoManager {
         return repository.fetchMemosByOwnerID(ownerID);
     }
 
-
     // Name - Memos
     public ArrayList<Memo> getMemosByNameAndOwnerID(String name, String ownerID){
         return repository.fetchMemosByNameAndOwnerID(name, ownerID);
+    }
+
+    public ArrayList<CalendarEvent> getEventsByMemoIDAndOwnerID(String memoID, String ownerID) {
+        return eventManager.getEventsByMemoIDAndOwnerID(memoID, ownerID);
     }
 
     // edit - Memos
@@ -55,56 +58,38 @@ public class MemoManager {
         return repository.editMemoNote(memoID, note, newNote, ownerID);
     }
 
-
-    boolean editMemoCountAdd(String memoID, String ownerID){
-        return repository.editMemoCountAdd(memoID, ownerID);
-    }
-
-    boolean editMemoCountRemove(String memoID, String ownerID){
-        return repository.editMemoCountRemove(memoID, ownerID);
-    }
-
-    // delete - Memo
-    boolean deleteMemo(String memoID, String ownerID){
-        return repository.deleteMemo(memoID, ownerID);
-    }
-
-    public ArrayList<CalendarEvent> fetchEventsByMemoIDAndOwnerID(String memoID, String ownerID){
-        ArrayList<CalendarEvent> events =  eventManager.getEventsByOwnerID(ownerID);
-        ArrayList<CalendarEvent> newEvents = new ArrayList<>();
-        for (CalendarEvent event : events){
-            for (String id : event.getMemoIDs()){
-                if (id.equals(memoID)){
-                    newEvents.add(event);
-                }
-            }
+    boolean deleteMemo(String memoID, String ownerID) {
+        Memo memo = repository.fetchMemoByMemoIDAndOwnerID(memoID, ownerID);
+        if (memo.getCount() <= 0) {
+            return repository.deleteMemo(memoID, ownerID);
         }
-        return newEvents;
+        return false;
     }
 
     // Add/remove memos
-    boolean addMemoToEvent(String memoID, String eventID, String ownerID){
-        CalendarEvent event = eventManager.getEventByIDAndUserID(eventID, ownerID);
-        ArrayList<String> ids = event.getTagIDs();
-        ids.add(memoID);
-        event.setTagIDs(ids);
-        getMemoByMemoID(memoID).addCount();
-        return true;
-    }
-    boolean removeMemoFromEvent(String memoID, String eventID, String ownerID){
-        CalendarEvent event = eventManager.getEventByIDAndUserID(eventID, ownerID);
-        ArrayList<String> ids = event.getMemoIDs();
-        ArrayList<String> newIDs = new ArrayList<>();
-        boolean removed = false;
-        for (String id : ids){
-            if (!id.equals(memoID)){
-                newIDs.add(id);
-            }else{
-                removed = true;
-                getMemoByMemoID(memoID).removeCount();
+    public boolean addMemoToEvent(String memoID, String eventID, String ownerID) {
+        Memo memo = repository.fetchMemoByMemoIDAndOwnerID(memoID, ownerID);
+        if (memo != null) {
+            boolean success = eventManager.editMemoID(eventID, memoID, ownerID);
+            if (success) {
+                repository.editMemoCountAdd(memoID, ownerID);
             }
+            return success;
         }
-        event.setMemoIDs(newIDs);
-        return removed;
+        return false;
+    }
+
+    public boolean removeMemoFromEvent(String eventID, String ownerID) {
+        CalendarEvent event = eventManager.getEventByIDAndUserID(eventID, ownerID);
+        if (event == null) {
+            return false;
+        }
+        String memoID = event.getMemoID();
+        if (memoID == "") {
+            return true;
+        }
+        eventManager.editMemoID(eventID, "", ownerID);
+        repository.editMemoCountRemove(memoID, ownerID);
+        return true;
     }
 }
